requirements.txt
pytorch_lightning
spacy
hydra-core
lightning
einops
pandas
opencv-python-headless==4.5.5.64
timm==0.4.12
fairscale
tabulate
transformers==4.33.2
salesforce-lavis
numpy==1.24.4
torch
lightning-utilities
# pip install git+https://github.com/openai/CLIP.git